+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## master -
+### Added
+- Method `getMonocularMatrix()` to get the intrinsic camera matrix of the monocular.
+- Method `getStereoLink()` to get the link between the stereo camera and the base frame.
 
 ## 1.0.2 - 2020-10-22
 ### Changed